Code optimization & fix project references (#1)
* Fix broken path references in project and solution

* Use == instead of Equals

If the object is already null and Equals is called, at runtime the program would still throw an NRE.

* Move post-command-execution tasks to CommandExecuted

See: https://docs.stillu.cc/

* Decrease code size by using LINQ

* Use : operator instead of Format text

* Revert "Use : operator instead of Format text"

This reverts commit 078656aed6862165d90e2a43f3c944f7e2307eb8.'

Removed file from project completely.

* Revert "Decrease code size by using LINQ "

This reverts commit 2626fa9f6b7e7035ad19f7537a7f947cba17e0fd.

Refactored a lot here to change storing the ranking info (level and job) in Dictionary to List, so I changed the code there since then. You are welcome to check it again if you would like and change it if you want. 👍 😄